+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v 0.2.0 
+- Allow publish templates from filebrowser 
+- Added loading icon when loading template list
+- Fixed a bug that failed to preview template when file name contains html special char
+
 # v 0.1.0
 - Support providing templates from different contents manager
 - Preview template before import it.